+++ packages/coin-eos/tests/eos.test.ts
@@ -49,6 +49,7 @@ describe("eos", () => {
     ps.push("L1v");
     ps.push("0x31342f041c5b54358074b4579231c8a300be65e687dff020bc7779598b428 97a");
     ps.push("0x31342f041c5b54358074b457。、。9231c8a300be65e687dff020bc7779598b428 97a");
+    ps.push("0000000000000000000000000000000000000000000000000000000000000000");
     test("edge test", async () => {
         const wallet = new EosWallet();
         let j = 1;
@@ -57,6 +58,7 @@ describe("eos", () => {
                 await wallet.getNewAddress({privateKey: ps[i]});
             } catch (e) {
                 j = j + 1
+                expect((await wallet.validPrivateKey({privateKey:ps[i]})).isValid).toEqual(false);
             }
         }
         expect(j).toEqual(ps.length+1);
@@ -64,8 +66,17 @@ describe("eos", () => {
     test("validPrivateKey", async () => {
         const wallet = new EosWallet();
         const privateKey = await wallet.getRandomPrivateKey();
-        const res = await wallet.validPrivateKey({privateKey:privateKey});
+        let res = await wallet.validPrivateKey({privateKey:privateKey});
         expect(res.isValid).toEqual(true);
+        res = await wallet.validPrivateKey({privateKey:privateKey+"1122"});
+        expect(res.isValid).toEqual(false);
+        res = await wallet.validPrivateKey({privateKey:""});
+        expect(res.isValid).toEqual(false);
+        res = await wallet.validPrivateKey({privateKey:privateKey.slice(1)});
+        expect(res.isValid).toEqual(false);
+
+        res = await wallet.validPrivateKey({privateKey:"0000000000000000000000000000000000000000000000000000000000000000"});
+        expect(res.isValid).toEqual(false);
     });
 
     test("address", async () => {