+++ packages/coin-kaspa/src/address.ts
@@ -6,7 +6,7 @@ export function pubKeyFromPrvKey(prvKey: string) {
         throw new Error("invalid key");
     }
     // todo check this rule
-    return base.toHex(signUtil.secp256k1.publicKeyCreate(base.fromHex(prvKey), true).slice(1));
+    return base.toHex(signUtil.secp256k1.publicKeyCreate(base.fromHex(prvKey.toLowerCase()), true).slice(1));
 }
 
 export function checkPrvKey(prvKey: string) {
@@ -14,7 +14,7 @@ export function checkPrvKey(prvKey: string) {
         return false;
     }
     const buf = base.fromHex(prvKey.toLowerCase());
-    return buf.length == 32
+    return buf.length == 32 && !buf.every(byte=>byte===0)
 }
 
 export function addressFromPubKey(pubKey: string) {