+++ packages/coin-base/CHANGELOG.md
@@ -2,7 +2,11 @@
 
 All notable changes to this project will be documented in this file.
 
+# [1.1.0](https://github.com/okx/js-wallet-sdk)(2024-01-15)
 
+### New Features
+
+- **coin-base:** support signCommonMsg ([](https://github.com/okx/js-wallet-sdk))
 
 # [1.0.14](https://github.com/okx/js-wallet-sdk) (2025-01-09)