+++ packages/coin-starknet/CHANGELOG.md
@@ -4,6 +4,12 @@
 All notable changes to this project will be documented in this file.
 
 
+# [1.1.0](https://github.com/okx/js-wallet-sdk)(2024-01-15)
+
+### New Features
+
+- **coin-starknet:** support signCommonMsg ([](https://github.com/okx/js-wallet-sdk))
+
 # [1.0.4](https://github.com/okx/js-wallet-sdk) (2024-08-20)
 
 ### Fix