+++ packages/coin-sui/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to this project will be documented in this file.
 
+# [1.1.0](https://github.com/okx/js-wallet-sdk)(2024-01-15)
+
+### New Features
+
+- **coin-sui:** support signCommonMsg ([](https://github.com/okx/js-wallet-sdk))
 
 # [1.0.5](https://github.com/okx/js-wallet-sdk) (2024-08-20)